Organizing the Store with Modules Namespacing Modules Accessing Namespaced Getters
