+++ build-a-bot/src/App.vue
@@ -41,7 +41,7 @@ export default {
   name: 'app',
   computed: {
     cart() {
-      return this.$store.state.cart;
+      return this.$store.state.robots.cart;
     },
   },
 };
