+++ build-a-bot/src/parts/get-parts-mixin.js
@@ -1,10 +1,11 @@
 export default {
   created() {
+    // eslint-disable-next-line no-undef
     this.$store.dispatch(getParts);
   },
   computed: {
     parts() {
-      return this.$store.state.parts || {
+      return this.$store.state.robots.parts || {
         heads: [],
         arms: [],
         torsos: [],
