+++ build-a-bot/src/store/index.js
@@ -1,38 +1,15 @@
 /* eslint-disable no-unused-vars */
 import Vue from 'vue';
 import Vuex from 'vuex';
-import axios from 'axios';
+
+import robotosModules from './modules/robots';
+import usersModules from './modules/users';
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
-  state: {
-    cart: [],
-    parts: null,
-  },
-  mutations: {
-    addRobotToCart(state, robot) {
-      state.cart.push(robot);
-    },
-    updateParts(state, parts) {
-      state.parts = parts;
-    },
-  },
-  actions: {
-    getParts({ commit }) {
-      axios.get('/api/parts')
-        .then(result => commit('updateParts', result.data))
-        .catch(console.error);
-    },
-    addRobotToCart({ commit, state }, robot) {
-      const cart = [...state.cart, robot];
-      axios.post('/api/cart', cart)
-        .then(() => commit('addRobotToCart', robot));
-    },
-  },
-  getters: {
-    cartSaleItems(state) {
-      return state.cart.filter(item => item.head.onSale);
-    },
+  modules: {
+    robots: robotosModules,
+    users: usersModules,
   },
 });
