+++ build-a-bot/src/store/modules/users.js
@@ -1,0 +1,24 @@
+import axios from 'axios';
+
+export default {
+  state: {
+    user: null,
+  },
+  mutations: {
+    updateCurrentUser(state, user) {
+      state.user = user;
+    },
+  },
+  getters: {
+  },
+  actions: {
+    signIn({ commit }) {
+      axios.post('/api/sign-in')
+        .then(result => commit('updateCurrentUser', result.data))
+        .catch(console.error);
+    },
+    addRobotToCart() {
+      console.log('Users addRobotToCart called');
+    },
+  },
+};
